notification/api/v1/serializers.py
from rest_framework.serializers import ModelSerializer

from manager_panel.models import Events


class UGCEventSerializer(ModelSerializer):
    class Meta:
        model = Events
        fields = "__all__"
