notification/requirements.txt
Django==3.2.13
django-cors-headers==3.11.0
django-debug-toolbar==3.4.0
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
Markdown==3.3.7
psycopg2-binary==2.9.3
python-dotenv==0.20.0